Do not use PVC is parallelism > 1.
EBS volumes can only be attached to one pod and I couldn't figure out a
way to mount the volume depending on the pod name.

An alternative would be to use efs but that provisioner is not available
everywhere and I don't want to spend the time (for now) to figure that
one out.

So for now if parallelism > 1 then retries will have to download all the
data again, it can be mitigated by doing small segments.